en proceso de pasar la interfaces a los componente
